requirements.txt
# =========================
# Cassandra
# =========================
cassandra-driver==3.29.2

# =========================
# Data processing
# =========================
pandas==2.2.2
numpy==1.26.4

# =========================
# Machine Learning / NLP
# =========================
scikit-learn==1.4.2
nltk==3.9.1

# =========================
# Visualization
# =========================
matplotlib==3.9.0
seaborn==0.13.2

# =========================
# Utils
# =========================
python-dotenv==1.0.1
tqdm==4.66.4
